backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
greenlet
alembic
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
email-validator
aiosqlite
pandas
openpyxl
thefuzz
python-levenshtein
aiohttp

